Refactor totalWork Calculation and Remove Redundant Method in SignatureModel
Create calculateTotalWork method in SignatureModel to centralise the calculation of the total workload for signature processes in comparison mode.

Adjust logic for initialising and populating trialsPerKeyByGroup array within calculateTotalWork and remove the calculateTrialsPerKeyByGroup method.

Update totalWork calculation to accurately reflect the workload in terms of operations needed for the cross-parameter benchmarking/comparison mode, including key groups and hash function considerations.